Initial int to long conversion work
ServiceStack library and methods have been updated to use longs where
appropriate.  Also added one failing test that should work with successful
long conversion.

Test code still needs to be updated.

+++ src/ServiceStack.Redis/BasicRedisClientManager.cs
@@ -36,7 +36,7 @@ namespace ServiceStack.Redis
 
 		public IRedisClientFactory RedisClientFactory { get; set; }
 
-		public int Db { get; private set; }
+		public long Db { get; private set; }
 
         public Action<IRedisNativeClient> ConnectionFilter { get; set; }
